+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 **v2.23.0**:
+- fixed issue with Drag&Drop of workspace roles
 - update sidebar icon to make it more distinctive
 - update [README](./README.md)
   - update Drag&Drop matrix